Drive-Github Merge
The project source folder is now hosted in a folder on Drive, which itself is the clone of this repository. the additional data was added via windows explorer, so that Github for Windows could commit the entire contents of the folder. Now, the repo clone directory itself is hosted on Drive. The same Drive folder is accessed by VS Code and RMMZ. This way, any changes made in VS Code are reflected immediately in RMMZ, and vice versa. These changes are made in Drive, enabling Github for Windows to commit changes to the repo.